Atualizada página de agendamento de São Paulo
